feat: line and polygon drawing

+++ src/gl/color.h
@@ -19,8 +19,8 @@ namespace color
         void setB(unsigned char b);
         friend std::ostream &operator<<(std::ostream &os, const Color &c); // |2| friend keyword
         friend Color operator+(const Color &lhs, const Color &rhs);        // |3| operator overloading
-        friend Color operator*(const Color &c, float f);                   // reminder -> use <num>f when using float XD
-        friend Color operator*(float f, const Color &c);
+        friend Color operator*(const Color &c, double f);
+        friend Color operator*(double f, const Color &c);
 
     private:
         unsigned char clamp(unsigned char value);
